tests/wpt/web-platform-tests/html/browsers/browsing-the-web/overlapping-navigations-and-traversals/tentative/cross-document-traversal-same-document-nav.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Same-document navigations during cross-document traversals</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<!--
  The spec currently says that same-document navigations must stop traverals,
  but this does not match browsers: https://github.com/whatwg/html/issues/6773.
  Browsers also disagree on how far back this should take us. This test assumes
  a behavior similar to Firefox's, although Firefox is inconsistent and only
  applies this behavior for fragments, not for pushState().
-->

<body>
<script type="module">
import { createIframe, waitForLoad, delay } from "./resources/helpers.mjs";

promise_test(async t => {
  const iframe = await createIframe(t);

  // Setup
  // Extra delay()s are necessary because if we navigate "inside" the load
  // handler (i.e. in a promise reaction for the load handler) then it will
  // be a replace navigation.
  iframe.contentWindow.location.search = "?1";
  await waitForLoad(iframe);
  await delay(t, 0);
  iframe.contentWindow.location.search = "?2";
  await waitForLoad(iframe);
  await delay(t, 0);

  iframe.contentWindow.history.back();

  assert_equals(iframe.contentWindow.location.search, "?2", "must not go back synchronously");

  iframe.contentWindow.location.hash = "#3";
  assert_equals(iframe.contentWindow.location.search, "?2");
  assert_equals(iframe.contentWindow.location.hash, "#3");

  // Eventually ends up on ?2
  await t.step_wait(() => iframe.contentWindow.location.search === "?2" && iframe.contentWindow.location.hash === "");
}, "same-document traversals are not canceled by fragment navigations");

promise_test(async t => {
  const iframe = await createIframe(t);

  // Setup
  // Extra delay()s are necessary because if we navigate "inside" the load
  // handler (i.e. in a promise reaction for the load handler) then it will
  // be a replace navigation.
  iframe.contentWindow.location.search = "?1";
  await waitForLoad(iframe);
  await delay(t, 0);
  iframe.contentWindow.location.search = "?2";
  await waitForLoad(iframe);
  await delay(t, 0);

  iframe.contentWindow.history.back();

  assert_equals(iframe.contentWindow.location.search, "?2", "must not go back synchronously");

  iframe.contentWindow.history.pushState(null, "", "?3");
  assert_equals(iframe.contentWindow.location.search, "?3");

  // Eventually ends up on ?2
  await t.step_wait(() => iframe.contentWindow.location.search === "?2");
}, "same-document traversals are not canceled by pushState()");
</script>
